More logging, and now I've found why postTaxFail doesn't work:
The setting of a channels log field is only done when calling get() when
creating a channel struct for the first time. After loading from JSON
however, this is not done, and hence name() can never resolve the
channel name, and so no posts are made to the parent bot.
Solution: maybe just change the Channel struct to include the name of
the channel, even if we then repeat it by having it as a parent map key.

This commit does not fix this. Only points to the problem.